manager/views.py
from rest_framework import generics
from .models import Artist, Event
from .serializers import ArtistSerializer, EventSerializer


class ArtistList(generics.ListCreateAPIView):
    queryset = Artist.objects.all()
    serializer_class = ArtistSerializer


class EventList(generics.ListCreateAPIView):
    queryset = Event.objects.all()
    serializer_class = EventSerializer
